requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
mediapipe>=0.10.0
Pillow>=10.0.0
python-dotenv>=1.0.0 